+++ extra/constants.js
@@ -1,0 +1,19 @@
+export const permissions = {
+    'getUsers': {
+        all: ['head-trainer'],
+        read : ['trainee', 'trainer'],
+        write : ['trainer'],
+        delete: [],
+    }
+}
+
+export const users = [
+    {
+        traineeEmail: 'trainee1@successive.tech',
+        reviewerEmail: 'reviewer1@successive.tech'
+    },
+    {
+        traineeEmail:'trainee2@successive.tech',
+        reviewerEmail:'reviewer2@gmail.com'
+    }
+]
