+++ extra/utils/helper.js
@@ -1,0 +1,6 @@
+let validateEmail = (email) =>{
+    const regex=/@successive.tech$/i;
+    return regex.test(email);
+}
+
+export default validateEmail;
