+++ extra/utils/index.js
@@ -1,0 +1,1 @@
+export { default as  hasPermission } from './permissions'
