+++ package.json
@@ -1,0 +1,12 @@
+{
+  "name": "freshers-training",
+  
+  "main": "index.js",
+  "devDependencies": {
+    "@babel/cli": "^7.15.7",
+    "@babel/core": "^7.15.8",
+    "@babel/node": "^7.15.8",
+    "@babel/preset-env": "^7.15.8"
+  }
+
+}
